+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [4.0.0] - 2026-04-03
 
 * Optional periodic import of history while listening for messages
+* Add support for Elasticsearch 9 in addition to Elasticsearch 8 (switch using `version` config property)
 * Switch to using uv as Python package manager
 * Use Alpine as base image for Docker image (decreases image size from over 900 MB down to around 180 MB)
 * Rename development branch from `master` to `main`